Реализовал 2-е задание.

+++ HW15_Serialization/CUITask2/Entities/Contacts.cs
@@ -2,11 +2,13 @@
 
 namespace CUITask2.Entities
 {
-    [XmlRoot(ElementName="CONTACTS")]
-    public class Contacts {
-        [XmlElement(ElementName="SmsPhoneNumber")]
-        public int SmsPhoneNumber { get; set; }
-        [XmlElement(ElementName="PhoneNeedActualize")]
+    [XmlRoot(ElementName = "CONTACTS")]
+    public class Contacts
+    {
+        [XmlElement(ElementName = "SmsPhoneNumber")]
+        public long SmsPhoneNumber { get; set; }
+
+        [XmlElement(ElementName = "PhoneNeedActualize")]
         public bool PhoneNeedActualize { get; set; }
     }
 }
