+++ HW15_Serialization/CUITask2/Entities/PostErrors.cs
@@ -2,11 +2,11 @@
 
 namespace CUITask2.Entities
 {
-    [XmlRoot(ElementName="POST_ERRORS")]
-    public class PostErrors {
-        [XmlElement(ElementName="All")]
-        public int All { get; set; }
-        [XmlElement(ElementName="Critical")]
-        public int Critical { get; set; }
+    [XmlRoot(ElementName = "POST_ERRORS")]
+    public class PostErrors
+    {
+        [XmlElement(ElementName = "All")] public long All { get; set; }
+
+        [XmlElement(ElementName = "Critical")] public long Critical { get; set; }
     }
 }
